+++ core/src/test/java/com/ibm/aem/aemadvancedrestrictions/core/restrictions/patterns/NodeNotExistsPatternTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2024 IBM iX
+ * Copyright 2024 - 2025 IBM iX
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
@@ -55,7 +55,7 @@ class NodeNotExistsPatternTest {
 
     @BeforeEach
     void setup() {
-        asset = UtilityFunctions.createAssetWithMetadataProperty("prop", "val");
+        asset = UtilityFunctions.createAssetWithMetadataPropertyAndRoot("prop", "val");
     }
 
     @Test