+++ core/src/test/java/com/ibm/aem/aemadvancedrestrictions/core/restrictions/patterns/NumberLessPatternTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2024 IBM iX
+ * Copyright 2024 - 2025 IBM iX
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
@@ -60,31 +60,31 @@ class NumberLessPatternTest {
 
     @Test
     void matches_noPropertyPresent() {
-        Tree asset = UtilityFunctions.createAssetWithMetadataProperty(PROP_NAME, null);
+        Tree asset = UtilityFunctions.createAssetWithMetadataPropertyAndRoot(PROP_NAME, null);
         assertFalse(pattern.matches(asset, null));
     }
 
     @Test
     void matches_propertyMatches() {
-        Tree asset = UtilityFunctions.createAssetWithMetadataProperty(PROP_NAME, 0L);
+        Tree asset = UtilityFunctions.createAssetWithMetadataPropertyAndRoot(PROP_NAME, 0L);
         assertTrue(pattern.matches(asset, null));
     }
 
     @Test
     void matches_propertyMatchesNot() {
-        Tree asset = UtilityFunctions.createAssetWithMetadataProperty(PROP_NAME, 2L);
+        Tree asset = UtilityFunctions.createAssetWithMetadataPropertyAndRoot(PROP_NAME, 2L);
         assertFalse(pattern.matches(asset, null));
     }
 
     @Test
     void matches_multiPropertyMatches() {
-        Tree asset = UtilityFunctions.createAssetWithMetadataProperty(PROP_NAME, new Long[] {0L, 2L});
+        Tree asset = UtilityFunctions.createAssetWithMetadataPropertyAndRoot(PROP_NAME, new Long[] {0L, 2L});
         assertTrue(pattern.matches(asset, null));
     }
 
     @Test
     void matches_multiPropertyMatchesNot() {
-        Tree asset = UtilityFunctions.createAssetWithMetadataProperty(PROP_NAME, new Long[] {2L, 3L});
+        Tree asset = UtilityFunctions.createAssetWithMetadataPropertyAndRoot(PROP_NAME, new Long[] {2L, 3L});
         assertFalse(pattern.matches(asset, null));
     }