+++ CHANGES.md
@@ -1,4 +1,8 @@
 # Release History
+* 2.0.0
+  * Allow property checks on Sling (ordered) folders
+  * Support hierarchical property checks on parent items
+
 * 1.0.2
   * Examples: removed WKND dependency